requirements.txt
flask
gunicorn
python-dotenv
cloudinary
psycopg2-binary
SQLAlchemy


